+++ components/restreamer/client/src/modals/RestreamModal.svelte
@@ -61,6 +61,12 @@
       if (!!current.id) {
         changed ||= current.withHls !== previous.withHls;
       }
+
+      if (!!current.id) {
+        changed ||=
+          current.withPlaybackEncoding !== previous.withPlaybackEncoding;
+      }
+
       submitable &&= changed;
     })
   );
@@ -71,6 +77,7 @@
     let variables = {
       key: restream.key,
       with_hls: restream.withHls,
+      with_playback_encoding: restream.withPlaybackEncoding,
     };
 
     if (restream.label) {
@@ -237,6 +244,18 @@
             /> with HLS endpoint</label
           >
         </div>
+
+        {#if $restreamStore.withPlayback()}
+          <div class="playback-encoding">
+            <label
+              ><input
+                class="uk-checkbox"
+                type="checkbox"
+                bind:checked={$restreamStore.withPlaybackEncoding}
+              /> with playback encoding</label
+            >
+          </div>
+        {/if}
 
         <div class="uk-section uk-section-xsmall backups-section">
           <button